lambda/ga_fetcher/requirements.txt
boto3==1.28.38
google-analytics-data==0.16.0
google-auth==2.23.0 